Resort! and some other bits
So, as time progressed I decided to go with a different name as
XmlArticleImport made little no sense at all any more. It can do CSV now
- as I was not quite pleased with the normal workflow of importing to
shopware with a csv file. I put in an instance of parsecsv (which is
freaking sweet!) and did some magic to import actual products. need to
figure out how to put in more info alongside the standard stuff like
ordernumber or name. ohyeah images seem to be quite extensive as well.