Crons n' stuff
So I incorperated a cronjob and set it up to be daily. Ofcourse this
just a setting in the shopware backend and can be changed. Also some
code has been moved into a dispatchhelper. Its kind of a stupid name,
but I couldnt think of anything better.
A delta import is also being made. Just need to figure out a more
elegant way to set import from directory XYZ and stuff like that.

yay beer btw.